chore: move discover_templates() to execute.py instead of create_session()
